cfg and minor edits

+++ src/Executors/Output/Genesis.Output.ProjectFiles/ProjectFilesGenerator.cs
@@ -36,9 +36,15 @@ namespace Genesis.Output.Poco
             {
                 var outPath = Path.Combine(proj.Value, proj.Key);
 
-                await File.WriteAllTextAsync(outPath.Replace('<', '_').Replace('>', '_'), Template.Raw); //hacky, can't save fileNames with '<' or '>' in the name
-
-                Text.White($"Wrote '"); Text.Yellow(outPath); Text.WhiteLine("'");
+                if (Config.Overwrite == "true")
+                {
+                    await File.WriteAllTextAsync(outPath.Replace('<', '_').Replace('>', '_'), Template.Raw); //hacky, can't save fileNames with '<' or '>' in the name
+                    Text.White($"Wrote '"); Text.Yellow(outPath); Text.WhiteLine("'");
+                }
+                else
+                {
+                    Text.White($"Skipped writing '"); Text.Yellow(outPath); Text.WhiteLine("' | overwrite = false");
+                }
             }
 
             var result = new OutputGenesisExecutionResult
